refactor: 컴포넌트 hook 적용
