+++ src/components/ComplexForm/index.tsx
@@ -1,42 +1,12 @@
-import { useState } from "react";
-import { renderLog } from "../utils";
-import { memo, useCallback } from "../@lib";
-import { useNotificationContext } from "../context/useNotificationContext";
+import { renderLog } from "../../utils";
+import { memo } from "../../@lib";
+import { useComplex } from "./useComplex";
 
 export const ComplexForm: React.FC = memo(() => {
   renderLog("ComplexForm rendered");
-  const { addNotification } = useNotificationContext();
-  const [formData, setFormData] = useState({
-    name: "",
-    email: "",
-    age: 0,
-    preferences: [] as string[],
-  });
 
-  const handleSubmit = useCallback(
-    (e: React.FormEvent) => {
-      e.preventDefault();
-      addNotification("폼이 성공적으로 제출되었습니다", "success");
-    },
-    [addNotification]
-  );
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setFormData((prev) => ({
-      ...prev,
-      [name]: name === "age" ? parseInt(value) || 0 : value,
-    }));
-  };
-
-  const handlePreferenceChange = (preference: string) => {
-    setFormData((prev) => ({
-      ...prev,
-      preferences: prev.preferences.includes(preference)
-        ? prev.preferences.filter((p) => p !== preference)
-        : [...prev.preferences, preference],
-    }));
-  };
+  const { formData, handleSubmit, handleInputChange, handlePreferenceChange } =
+    useComplex();
 
   return (
     <div className="mt-8">
