+++ src/components/NotificationSystem/index.tsx
@@ -1,6 +1,6 @@
-import { memo } from "../@lib";
-import { useNotificationContext } from "../context/useNotificationContext";
-import { renderLog } from "../utils";
+import { memo } from "../../@lib";
+import { useNotificationContext } from "../../context/useNotificationContext";
+import { renderLog } from "../../utils";
 
 export const NotificationSystem: React.FC = memo(() => {
   renderLog("NotificationSystem rendered");
